requirement.txt
chainlit
langchain-community
langchain
chromadb
sentence-transformers
google-generativeai
python-dotenv
requests
fastapi
uvicorn
sqlite3
